requirements.txt
faster-whisper>=1.0.0
opencv-python>=4.8.0
pyzbar>=0.1.9
python-docx>=1.1.0
requests>=2.31.0
flask>=3.0.0
sounddevice>=0.4.6
numpy>=1.24.0
pycaw>=20230407
comtypes>=1.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
